# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Recette API",
    "description": "API pour gérer les recettes et les livres",
    "contact": {
      "name": "Support API",
      "email": "contact@example.com"
    }
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
    "/getRecette": {
      "get": {
        "tags": ["Recette"],
        "summary": "Récupérer toutes les recettes",
        "responses": {
          "200": {
            "description": "Liste des recettes récupérée avec succès",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Recette"
              }
            }
          },
          "500": {
            "$ref": "#/definitions/errorResponse.500"
          }
        }
      }
    },
    "/postRecette": {
      "post": {
        "tags": ["Recette"],
        "summary": "Ajouter une nouvelle recette",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Recette"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Recette ajoutée avec succès",
            "schema": {
              "$ref": "#/definitions/Recette"
            }
          },
          "400": {
            "$ref": "#/definitions/errorResponse.400"
          }
        }
      }
    },
    "/putRecette/{id}": {
      "put": {
        "tags": ["Recette"],
        "summary": "Mettre à jour une recette",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Recette"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Recette mise à jour avec succès",
            "schema": {
              "$ref": "#/definitions/Recette"
            }
          },
          "404": {
            "$ref": "#/definitions/errorResponse.404"
          }
        }
      }
    },
    "/deleteRecette/{id}": {
      "delete": {
        "tags": ["Recette"],
        "summary": "Supprimer une recette",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Recette supprimée avec succès"
          },
          "404": {
            "$ref": "#/definitions/errorResponse.404"
          }
        }
      }
    },
    "/getLivre": {
      "get": {
        "tags": ["Livre"],
        "summary": "Récupérer tous les livres",
        "responses": {
          "200": {
            "description": "Liste des livres récupérée avec succès",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Livre"
              }
            }
          },
          "500": {
            "$ref": "#/definitions/errorResponse.500"
          }
        }
      }
    },
    "/postLivre": {
      "post": {
        "tags": ["Livre"],
        "summary": "Ajouter un nouveau livre",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Livre"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Livre ajouté avec succès",
            "schema": {
              "$ref": "#/definitions/Livre"
            }
          },
          "400": {
            "$ref": "#/definitions/errorResponse.400"
          }
        }
      }
    },
    "/putLivre/{id}": {
      "put": {
        "tags": ["Livre"],
        "summary": "Mettre à jour un livre",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Livre"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Livre mis à jour avec succès",
            "schema": {
              "$ref": "#/definitions/Livre"
            }
          },
          "404": {
            "$ref": "#/definitions/errorResponse.404"
          }
        }
      }
    },
    "/deleteLivre/{id}": {
      "delete": {
        "tags": ["Livre"],
        "summary": "Supprimer un livre",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Livre supprimé avec succès"
          },
          "404": {
            "$ref": "#/definitions/errorResponse.404"
          }
        }
      }
    }
  },
  "definitions": {
    "Recette": {
      "type": "object",
      "properties": {
        "id": { "type": "string", "example": "12345" },
        "Titre": { "type": "string", "example": "Tarte aux pommes" },
        "Recette": { "type": "string", "example": "Mélangez les ingrédients et faites cuire." }
      }
    },
    "Livre": {
      "type": "object",
      "properties": {
        "id": { "type": "string", "example": "67890" },
        "Titre": { "type": "string", "example": "Livre de recettes" },
        "Description": { "type": "string", "example": "Un livre contenant diverses recettes de gâteaux." }
      }
    },
    "errorResponse.400": {
      "type": "object",
      "properties": {
        "code": { "type": "number", "example": 400 },
        "message": { "type": "string", "example": "Requête invalide." }
      }
    },
    "errorResponse.404": {
      "type": "object",
      "properties": {
        "code": { "type": "number", "example": 404 },
        "message": { "type": "string", "example": "Ressource non trouvée." }
      }
    },
    "errorResponse.500": {
      "type": "object",
      "properties": {
        "code": { "type": "number", "example": 500 },
        "message": { "type": "string", "example": "Erreur serveur." }
      }
    }
  }
}